+++ wow_login_messages/src/logon/version_3/cmd_auth_logon_proof_client.rs
@@ -59,7 +59,7 @@ impl CMD_AUTH_LOGON_PROOF_Client {
         }
 
         // security_flag: SecurityFlag
-        w.write_all(&u8::from(self.security_flag.as_int()).to_le_bytes())?;
+        w.write_all(&(self.security_flag.as_int().to_le_bytes()))?;
 
         match &self.security_flag {
             CMD_AUTH_LOGON_PROOF_Client_SecurityFlag::Pin {
