+++ wow_login_messages/src/logon/version_8/cmd_auth_logon_challenge_server.rs
@@ -64,7 +64,7 @@ impl CMD_AUTH_LOGON_CHALLENGE_Server {
         w.write_all(&Self::PROTOCOL_VERSION_VALUE.to_le_bytes())?;
 
         // result: LoginResult
-        w.write_all(&u8::from(self.result.as_int()).to_le_bytes())?;
+        w.write_all(&(self.result.as_int().to_le_bytes()))?;
 
         match &self.result {
             CMD_AUTH_LOGON_CHALLENGE_Server_LoginResult::Success {
@@ -107,7 +107,7 @@ impl CMD_AUTH_LOGON_CHALLENGE_Server {
                 }
 
                 // security_flag: SecurityFlag
-                w.write_all(&u8::from(security_flag.as_int()).to_le_bytes())?;
+                w.write_all(&(security_flag.as_int().to_le_bytes()))?;
 
                 if let Some(if_statement) = &security_flag.pin {
                     // pin_grid_seed: u32
