+++ wow_login_messages/src/logon/version_8/cmd_auth_reconnect_proof_server.rs
@@ -37,7 +37,7 @@ impl CMD_AUTH_RECONNECT_PROOF_Server {
         w.write_all(&Self::OPCODE.to_le_bytes())?;
 
         // result: LoginResult
-        w.write_all(&u8::from(self.result.as_int()).to_le_bytes())?;
+        w.write_all(&(self.result.as_int().to_le_bytes()))?;
 
         // padding: u16
         w.write_all(&Self::PADDING_VALUE.to_le_bytes())?;
