+++ wow_login_messages/src/logon/version_8/realm.rs
@@ -41,13 +41,13 @@ pub struct Realm {
 impl Realm {
     pub(crate) fn write_into_vec(&self, mut w: impl std::io::Write) -> Result<(), std::io::Error> {
         // realm_type: RealmType
-        w.write_all(&u8::from(self.realm_type.as_int()).to_le_bytes())?;
+        w.write_all(&(self.realm_type.as_int().to_le_bytes()))?;
 
         // locked: Bool
         w.write_all(u8::from(self.locked).to_le_bytes().as_slice())?;
 
         // flag: RealmFlag
-        w.write_all(&u8::from(self.flag.as_int()).to_le_bytes())?;
+        w.write_all(&(self.flag.as_int().to_le_bytes()))?;
 
         // name: CString
         // TODO: Guard against strings that are already null-terminated
@@ -64,13 +64,13 @@ impl Realm {
         w.write_all(&[0])?;
 
         // population: Population
-        w.write_all(&u32::from(self.population.as_int()).to_le_bytes())?;
+        w.write_all(&(self.population.as_int().to_le_bytes()))?;
 
         // number_of_characters_on_realm: u8
         w.write_all(&self.number_of_characters_on_realm.to_le_bytes())?;
 
         // category: RealmCategory
-        w.write_all(&u8::from(self.category.as_int()).to_le_bytes())?;
+        w.write_all(&(self.category.as_int().to_le_bytes()))?;
 
         // realm_id: u8
         w.write_all(&self.realm_id.to_le_bytes())?;
