+++ wow_world_messages/src/world/shared/cmsg_arena_team_remove_tbc_wrath.rs
@@ -36,7 +36,7 @@ impl crate::Message for CMSG_ARENA_TEAM_REMOVE {
     }
     fn read_body<S: crate::private::Sealed>(mut r: &mut &[u8], body_size: u32) -> std::result::Result<Self, crate::errors::ParseError> {
         if !(5..=260).contains(&body_size) {
-            return Err(crate::errors::ParseError::InvalidSize { opcode: 0x0354, size: body_size as u32 });
+            return Err(crate::errors::ParseError::InvalidSize { opcode: 0x0354, size: body_size });
         }
 
         // arena_team: u32
