+++ wow_world_messages/src/world/shared/cmsg_auction_list_bidder_items_vanilla_tbc_wrath.rs
@@ -45,7 +45,7 @@ impl crate::Message for CMSG_AUCTION_LIST_BIDDER_ITEMS {
     }
     fn read_body<S: crate::private::Sealed>(mut r: &mut &[u8], body_size: u32) -> std::result::Result<Self, crate::errors::ParseError> {
         if !(16..=10240).contains(&body_size) {
-            return Err(crate::errors::ParseError::InvalidSize { opcode: 0x0264, size: body_size as u32 });
+            return Err(crate::errors::ParseError::InvalidSize { opcode: 0x0264, size: body_size });
         }
 
         // auctioneer: Guid
