+++ wow_world_messages/src/world/shared/cmsg_auth_session_vanilla_tbc.rs
@@ -75,7 +75,7 @@ impl crate::Message for CMSG_AUTH_SESSION {
     }
     fn read_body<S: crate::private::Sealed>(mut r: &mut &[u8], body_size: u32) -> std::result::Result<Self, crate::errors::ParseError> {
         if !(37..=65827).contains(&body_size) {
-            return Err(crate::errors::ParseError::InvalidSize { opcode: 0x01ED, size: body_size as u32 });
+            return Err(crate::errors::ParseError::InvalidSize { opcode: 0x01ED, size: body_size });
         }
 
         // build: u32
