+++ wow_world_messages/src/world/shared/cmsg_autobank_item_vanilla_tbc_wrath.rs
@@ -32,7 +32,7 @@ impl crate::Message for CMSG_AUTOBANK_ITEM {
     }
     fn read_body<S: crate::private::Sealed>(mut r: &mut &[u8], body_size: u32) -> std::result::Result<Self, crate::errors::ParseError> {
         if body_size != 2 {
-            return Err(crate::errors::ParseError::InvalidSize { opcode: 0x0283, size: body_size as u32 });
+            return Err(crate::errors::ParseError::InvalidSize { opcode: 0x0283, size: body_size });
         }
 
         // bag_index: u8
