+++ wow_world_messages/src/world/shared/cmsg_autostore_bag_item_vanilla_tbc_wrath.rs
@@ -37,7 +37,7 @@ impl crate::Message for CMSG_AUTOSTORE_BAG_ITEM {
     }
     fn read_body<S: crate::private::Sealed>(mut r: &mut &[u8], body_size: u32) -> std::result::Result<Self, crate::errors::ParseError> {
         if body_size != 3 {
-            return Err(crate::errors::ParseError::InvalidSize { opcode: 0x010B, size: body_size as u32 });
+            return Err(crate::errors::ParseError::InvalidSize { opcode: 0x010B, size: body_size });
         }
 
         // source_bag: u8
