+++ wow_world_messages/src/world/shared/cmsg_battlemaster_join_arena_tbc_wrath.rs
@@ -33,7 +33,7 @@ impl crate::Message for CMSG_BATTLEMASTER_JOIN_ARENA {
         w.write_all(&self.battlemaster.guid().to_le_bytes())?;
 
         // arena_type: JoinArenaType
-        w.write_all(&u8::from(self.arena_type.as_int()).to_le_bytes())?;
+        w.write_all(&(self.arena_type.as_int().to_le_bytes()))?;
 
         // as_group: Bool
         w.write_all(u8::from(self.as_group).to_le_bytes().as_slice())?;
@@ -45,7 +45,7 @@ impl crate::Message for CMSG_BATTLEMASTER_JOIN_ARENA {
     }
     fn read_body<S: crate::private::Sealed>(mut r: &mut &[u8], body_size: u32) -> std::result::Result<Self, crate::errors::ParseError> {
         if body_size != 11 {
-            return Err(crate::errors::ParseError::InvalidSize { opcode: 0x0358, size: body_size as u32 });
+            return Err(crate::errors::ParseError::InvalidSize { opcode: 0x0358, size: body_size });
         }
 
         // battlemaster: Guid
