+++ wow_world_messages/src/world/shared/cmsg_buy_item_vanilla_tbc.rs
@@ -46,7 +46,7 @@ impl crate::Message for CMSG_BUY_ITEM {
     }
     fn read_body<S: crate::private::Sealed>(mut r: &mut &[u8], body_size: u32) -> std::result::Result<Self, crate::errors::ParseError> {
         if body_size != 14 {
-            return Err(crate::errors::ParseError::InvalidSize { opcode: 0x01A2, size: body_size as u32 });
+            return Err(crate::errors::ParseError::InvalidSize { opcode: 0x01A2, size: body_size });
         }
 
         // vendor: Guid
