+++ wow_world_messages/src/world/shared/cmsg_clear_trade_item_vanilla_tbc_wrath.rs
@@ -27,7 +27,7 @@ impl crate::Message for CMSG_CLEAR_TRADE_ITEM {
     }
     fn read_body<S: crate::private::Sealed>(mut r: &mut &[u8], body_size: u32) -> std::result::Result<Self, crate::errors::ParseError> {
         if body_size != 1 {
-            return Err(crate::errors::ParseError::InvalidSize { opcode: 0x011E, size: body_size as u32 });
+            return Err(crate::errors::ParseError::InvalidSize { opcode: 0x011E, size: body_size });
         }
 
         // trade_slot: u8
