+++ wow_world_messages/src/world/shared/cmsg_dblookup_vanilla_tbc_wrath.rs
@@ -35,7 +35,7 @@ impl crate::Message for CMSG_DBLOOKUP {
     }
     fn read_body<S: crate::private::Sealed>(mut r: &mut &[u8], body_size: u32) -> std::result::Result<Self, crate::errors::ParseError> {
         if !(1..=256).contains(&body_size) {
-            return Err(crate::errors::ParseError::InvalidSize { opcode: 0x0002, size: body_size as u32 });
+            return Err(crate::errors::ParseError::InvalidSize { opcode: 0x0002, size: body_size });
         }
 
         // query: CString
