+++ wow_world_messages/src/world/shared/cmsg_destroyitem_vanilla_tbc_wrath.rs
@@ -52,7 +52,7 @@ impl crate::Message for CMSG_DESTROYITEM {
     }
     fn read_body<S: crate::private::Sealed>(mut r: &mut &[u8], body_size: u32) -> std::result::Result<Self, crate::errors::ParseError> {
         if body_size != 6 {
-            return Err(crate::errors::ParseError::InvalidSize { opcode: 0x0111, size: body_size as u32 });
+            return Err(crate::errors::ParseError::InvalidSize { opcode: 0x0111, size: body_size });
         }
 
         // bag: u8
