+++ wow_world_messages/src/world/shared/cmsg_duel_accepted_vanilla_tbc_wrath.rs
@@ -29,7 +29,7 @@ impl crate::Message for CMSG_DUEL_ACCEPTED {
     }
     fn read_body<S: crate::private::Sealed>(mut r: &mut &[u8], body_size: u32) -> std::result::Result<Self, crate::errors::ParseError> {
         if body_size != 8 {
-            return Err(crate::errors::ParseError::InvalidSize { opcode: 0x016C, size: body_size as u32 });
+            return Err(crate::errors::ParseError::InvalidSize { opcode: 0x016C, size: body_size });
         }
 
         // guid: Guid
