+++ wow_world_messages/src/world/shared/cmsg_far_sight_vanilla_tbc_wrath.rs
@@ -23,13 +23,13 @@ impl crate::Message for CMSG_FAR_SIGHT {
 
     fn write_into_vec(&self, mut w: impl std::io::Write) -> Result<(), std::io::Error> {
         // operation: FarSightOperation
-        w.write_all(&u8::from(self.operation.as_int()).to_le_bytes())?;
+        w.write_all(&(self.operation.as_int().to_le_bytes()))?;
 
         Ok(())
     }
     fn read_body<S: crate::private::Sealed>(mut r: &mut &[u8], body_size: u32) -> std::result::Result<Self, crate::errors::ParseError> {
         if body_size != 1 {
-            return Err(crate::errors::ParseError::InvalidSize { opcode: 0x027A, size: body_size as u32 });
+            return Err(crate::errors::ParseError::InvalidSize { opcode: 0x027A, size: body_size });
         }
 
         // operation: FarSightOperation
