+++ wow_world_messages/src/world/shared/cmsg_gmsurvey_submit_vanilla_tbc_wrath.rs
@@ -50,7 +50,7 @@ impl crate::Message for CMSG_GMSURVEY_SUBMIT {
     }
     fn read_body<S: crate::private::Sealed>(mut r: &mut &[u8], body_size: u32) -> std::result::Result<Self, crate::errors::ParseError> {
         if !(65..=2870).contains(&body_size) {
-            return Err(crate::errors::ParseError::InvalidSize { opcode: 0x032A, size: body_size as u32 });
+            return Err(crate::errors::ParseError::InvalidSize { opcode: 0x032A, size: body_size });
         }
 
         // survey_id: u32
