+++ wow_world_messages/src/world/shared/cmsg_gmticket_updatetext_tbc_wrath.rs
@@ -33,7 +33,7 @@ impl crate::Message for CMSG_GMTICKET_UPDATETEXT {
     }
     fn read_body<S: crate::private::Sealed>(mut r: &mut &[u8], body_size: u32) -> std::result::Result<Self, crate::errors::ParseError> {
         if !(1..=256).contains(&body_size) {
-            return Err(crate::errors::ParseError::InvalidSize { opcode: 0x0207, size: body_size as u32 });
+            return Err(crate::errors::ParseError::InvalidSize { opcode: 0x0207, size: body_size });
         }
 
         // message: CString
