+++ wow_world_messages/src/world/shared/cmsg_group_assistant_leader_vanilla_tbc_wrath.rs
@@ -34,7 +34,7 @@ impl crate::Message for CMSG_GROUP_ASSISTANT_LEADER {
     }
     fn read_body<S: crate::private::Sealed>(mut r: &mut &[u8], body_size: u32) -> std::result::Result<Self, crate::errors::ParseError> {
         if body_size != 9 {
-            return Err(crate::errors::ParseError::InvalidSize { opcode: 0x028F, size: body_size as u32 });
+            return Err(crate::errors::ParseError::InvalidSize { opcode: 0x028F, size: body_size });
         }
 
         // guid: Guid
