+++ wow_world_messages/src/world/shared/cmsg_guild_rank_tbc_wrath.rs
@@ -45,7 +45,7 @@ impl crate::Message for CMSG_GUILD_RANK {
         w.write_all(&[0])?;
 
         // money_per_day: Gold
-        w.write_all(u32::from(self.money_per_day.as_int()).to_le_bytes().as_slice())?;
+        w.write_all((self.money_per_day.as_int()).to_le_bytes().as_slice())?;
 
         // bank_tab_rights: GuildBankRights[6]
         for i in self.bank_tab_rights.iter() {
@@ -56,7 +56,7 @@ impl crate::Message for CMSG_GUILD_RANK {
     }
     fn read_body<S: crate::private::Sealed>(mut r: &mut &[u8], body_size: u32) -> std::result::Result<Self, crate::errors::ParseError> {
         if !(61..=316).contains(&body_size) {
-            return Err(crate::errors::ParseError::InvalidSize { opcode: 0x0231, size: body_size as u32 });
+            return Err(crate::errors::ParseError::InvalidSize { opcode: 0x0231, size: body_size });
         }
 
         // rank_id: u32
