+++ wow_world_messages/src/world/shared/cmsg_item_name_query_vanilla_tbc_wrath.rs
@@ -34,7 +34,7 @@ impl crate::Message for CMSG_ITEM_NAME_QUERY {
     }
     fn read_body<S: crate::private::Sealed>(mut r: &mut &[u8], body_size: u32) -> std::result::Result<Self, crate::errors::ParseError> {
         if body_size != 12 {
-            return Err(crate::errors::ParseError::InvalidSize { opcode: 0x02C4, size: body_size as u32 });
+            return Err(crate::errors::ParseError::InvalidSize { opcode: 0x02C4, size: body_size });
         }
 
         // item: u32
