+++ wow_world_messages/src/world/shared/cmsg_item_text_query_vanilla_tbc.rs
@@ -41,7 +41,7 @@ impl crate::Message for CMSG_ITEM_TEXT_QUERY {
     }
     fn read_body<S: crate::private::Sealed>(mut r: &mut &[u8], body_size: u32) -> std::result::Result<Self, crate::errors::ParseError> {
         if body_size != 12 {
-            return Err(crate::errors::ParseError::InvalidSize { opcode: 0x0243, size: body_size as u32 });
+            return Err(crate::errors::ParseError::InvalidSize { opcode: 0x0243, size: body_size });
         }
 
         // item_text_id: u32
