+++ wow_world_messages/src/world/shared/cmsg_join_channel_tbc_wrath.rs
@@ -55,7 +55,7 @@ impl crate::Message for CMSG_JOIN_CHANNEL {
     }
     fn read_body<S: crate::private::Sealed>(mut r: &mut &[u8], body_size: u32) -> std::result::Result<Self, crate::errors::ParseError> {
         if !(8..=518).contains(&body_size) {
-            return Err(crate::errors::ParseError::InvalidSize { opcode: 0x0097, size: body_size as u32 });
+            return Err(crate::errors::ParseError::InvalidSize { opcode: 0x0097, size: body_size });
         }
 
         // channel_id: u32
