+++ wow_world_messages/src/world/shared/cmsg_loot_master_give_vanilla_tbc_wrath.rs
@@ -39,7 +39,7 @@ impl crate::Message for CMSG_LOOT_MASTER_GIVE {
     }
     fn read_body<S: crate::private::Sealed>(mut r: &mut &[u8], body_size: u32) -> std::result::Result<Self, crate::errors::ParseError> {
         if body_size != 17 {
-            return Err(crate::errors::ParseError::InvalidSize { opcode: 0x02A3, size: body_size as u32 });
+            return Err(crate::errors::ParseError::InvalidSize { opcode: 0x02A3, size: body_size });
         }
 
         // loot: Guid
