+++ wow_world_messages/src/world/shared/cmsg_loot_roll_tbc_wrath.rs
@@ -34,13 +34,13 @@ impl crate::Message for CMSG_LOOT_ROLL {
         w.write_all(&self.item_slot.to_le_bytes())?;
 
         // vote: RollVote
-        w.write_all(&u8::from(self.vote.as_int()).to_le_bytes())?;
+        w.write_all(&(self.vote.as_int().to_le_bytes()))?;
 
         Ok(())
     }
     fn read_body<S: crate::private::Sealed>(mut r: &mut &[u8], body_size: u32) -> std::result::Result<Self, crate::errors::ParseError> {
         if body_size != 13 {
-            return Err(crate::errors::ParseError::InvalidSize { opcode: 0x02A0, size: body_size as u32 });
+            return Err(crate::errors::ParseError::InvalidSize { opcode: 0x02A0, size: body_size });
         }
 
         // item: Guid
