+++ wow_world_messages/src/world/shared/cmsg_mail_take_item_tbc_wrath.rs
@@ -39,7 +39,7 @@ impl crate::Message for CMSG_MAIL_TAKE_ITEM {
     }
     fn read_body<S: crate::private::Sealed>(mut r: &mut &[u8], body_size: u32) -> std::result::Result<Self, crate::errors::ParseError> {
         if body_size != 16 {
-            return Err(crate::errors::ParseError::InvalidSize { opcode: 0x0246, size: body_size as u32 });
+            return Err(crate::errors::ParseError::InvalidSize { opcode: 0x0246, size: body_size });
         }
 
         // mailbox: Guid
