+++ wow_world_messages/src/world/shared/cmsg_meetingstone_info_vanilla_tbc.rs
@@ -22,7 +22,7 @@ impl crate::Message for CMSG_MEETINGSTONE_INFO {
     }
     fn read_body<S: crate::private::Sealed>(mut r: &mut &[u8], body_size: u32) -> std::result::Result<Self, crate::errors::ParseError> {
         if body_size != 0 {
-            return Err(crate::errors::ParseError::InvalidSize { opcode: 0x0296, size: body_size as u32 });
+            return Err(crate::errors::ParseError::InvalidSize { opcode: 0x0296, size: body_size });
         }
 
         Ok(Self {
