+++ wow_world_messages/src/world/shared/cmsg_move_set_raw_position_vanilla_tbc_wrath.rs
@@ -36,7 +36,7 @@ impl crate::Message for CMSG_MOVE_SET_RAW_POSITION {
     }
     fn read_body<S: crate::private::Sealed>(mut r: &mut &[u8], body_size: u32) -> std::result::Result<Self, crate::errors::ParseError> {
         if body_size != 16 {
-            return Err(crate::errors::ParseError::InvalidSize { opcode: 0x00E1, size: body_size as u32 });
+            return Err(crate::errors::ParseError::InvalidSize { opcode: 0x00E1, size: body_size });
         }
 
         // position: Vector3d
