+++ wow_world_messages/src/world/shared/cmsg_next_cinematic_camera_vanilla_tbc_wrath.rs
@@ -24,7 +24,7 @@ impl crate::Message for CMSG_NEXT_CINEMATIC_CAMERA {
     }
     fn read_body<S: crate::private::Sealed>(mut r: &mut &[u8], body_size: u32) -> std::result::Result<Self, crate::errors::ParseError> {
         if body_size != 0 {
-            return Err(crate::errors::ParseError::InvalidSize { opcode: 0x00FB, size: body_size as u32 });
+            return Err(crate::errors::ParseError::InvalidSize { opcode: 0x00FB, size: body_size });
         }
 
         Ok(Self {
